# public/openapi.yaml
...
openapi: 3.0.3
info:
  title: AccessGate API
  version: 0.1.0
  description: Event management and ticketing system API
  contact:
    name: API Support
    email: azeezwaris17@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: Development server
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Admin login
      description: Authenticate admin with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "admin@example.com"
              password: "password123"
      responses:
        '200':
          description: Successfully logged in
          content:
            application/json:
              example:
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                admin:
                  id: "507f1f77bcf86cd799439011"
                  fullName: "John Doe"
                  email: "admin@example.com"
        '400':
          description: Missing required fields
        '401':
          description: Invalid credentials
        '500':
          description: Internal server error
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new admin
      description: Register a new admin using a registration key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              fullName: "Jane Smith"
              email: "jane@example.com"
              password: "password123"
              registrationKey: "ABCD-1234-EFGH-5678"
      responses:
        '200':
          description: Successfully registered
          content:
            application/json:
              example:
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                admin:
                  id: "507f1f77bcf86cd799439012"
                  fullName: "Jane Smith"
                  email: "jane@example.com"
        '400':
          description: Missing fields or invalid registration key format
        '500':
          description: Internal server error

  # ==================== ADMIN MANAGEMENT ENDPOINTS ====================
  /admin/keys:
    get:
      tags:
        - Admin Management
      summary: Get all registration keys
      description: Retrieve all registration keys for the authenticated admin
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successfully retrieved registration keys
          content:
            application/json:
              example:
                keys:
                  - key: "ABCD-1234-EFGH-5678"
                    createdAt: "2024-01-15T10:30:00Z"
        '401':
          description: Unauthorized - Invalid or missing token
        '404':
          description: Admin not found
        '500':
          description: Internal server error
    post:
      tags:
        - Admin Management
      summary: Generate new registration key
      description: Generate a new registration key for the authenticated admin
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successfully generated registration key
          content:
            application/json:
              example:
                key: "ABCD-1234-EFGH-5678"
                allKeys:
                  - key: "ABCD-1234-EFGH-5678"
                    createdAt: "2024-01-15T10:30:00Z"
        '401':
          description: Unauthorized - Invalid or missing token
        '404':
          description: Admin not found
        '500':
          description: Internal server error

  # ==================== EVENT MANAGEMENT ENDPOINTS ====================
  /events:
    get:
      tags:
        - Events
      summary: Get all events with filtering and pagination
      description: Retrieve all events for the authenticated admin with optional filtering and pagination
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, active, completed, cancelled]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 10
      responses:
        '200':
          description: Successfully retrieved events
          content:
            application/json:
              example:
                success: true
                events: []
                pagination:
                  currentPage: 1
                  totalPages: 1
                  totalEvents: 0
                  hasNext: false
                  hasPrev: false
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
...
  /events/{eventId}:
    get:
      tags:
        - Event Operations
      summary: Get event details
      description: Retrieve detailed information for a specific event
      security:
        - BearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved event details
          content:
            application/json:
              example:
                success: true
                event:
                  _id: "507f1f77bcf86cd799439013"
                  name: "Tech Conference 2024"
                  date: "2024-06-15"
                  time: "09:00"
                  location: "Convention Center"
                  status: "active"
                  totalRevenue: 1000.00
                  totalTickets: 200
                  totalSold: 10
                  totalAvailable: 190
        '401':
          description: Unauthorized
        '404':
          description: Event not found or access denied
        '500':
          description: Internal server error
  /events/{eventId}/create-ticket:
    post:
      tags:
        - Event Operations
      summary: Create ticket for event
      description: Create a new ticket for a specific event
      security:
        - BearerAuth: []
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketRequest'
            example:
              eventId: "507f1f77bcf86cd799439013"
              ticketTypeId: "507f1f77bcf86cd799439014"
              attendeeName: "John Doe"
              attendeeEmail: "john@example.com"
      responses:
        '201':
          description: Successfully created ticket
          content:
            application/json:
              example:
                _id: "507f1f77bcf86cd799439015"
                eventId: "507f1f77bcf86cd799439013"
                ticketTypeId: "507f1f77bcf86cd799439014"
                attendeeName: "John Doe"
                attendeeEmail: "john@example.com"
                ticketCode: "AG-1705314600000-1A2B3C4D"
                status: "valid"
        '400':
          description: Missing required fields
        '401':
          description: Unauthorized
        '404':
          description: Event not found
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RegisterRequest:
      type: object
      required:
        - fullName
        - email
        - password
        - registrationKey
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        registrationKey:
          type: string
          pattern: "^[A-Z0-9]{4}-[A-Z0-9]{4}-[A-Z0-9]{4}-[A-Z0-9]{4}$"
...
    CreateTicketRequest:
      type: object
      required:
        - eventId
        - ticketTypeId
        - attendeeName
        - attendeeEmail
      properties:
        eventId:
          type: string
        ticketTypeId:
          type: string
        attendeeName:
          type: string
        attendeeEmail:
          type: string
          format: email
...